+++ src/libraries/DAQ/fa250Mode1Hit_factory.cc
@@ -24,7 +24,7 @@ jerror_t fa250Mode1Hit_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t fa250Mode1Hit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
+jerror_t fa250Mode1Hit_factory::brun(jana::JEventLoop *eventLoop, int32_t runnumber)
 {
 	return NOERROR;
 }
@@ -32,7 +32,7 @@ jerror_t fa250Mode1Hit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t fa250Mode1Hit_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t fa250Mode1Hit_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 
 	// Code to generate factory data goes here. Add it like:
