+++ src/libraries/system/BDXEventProcessor.h
@@ -39,8 +39,8 @@ class BDXEventProcessor:public JEventProcessor
 
 	private:
 		jerror_t init();                                 // Called once at program start.
-		jerror_t brun(JEventLoop*, int runnumber);       // Called everytime a new run number is detected.
-		jerror_t evnt(JEventLoop*, int eventnumber);     // Called every event.
+		jerror_t brun(JEventLoop*, int32_t runnumber);       // Called everytime a new run number is detected.
+		jerror_t evnt(JEventLoop*, uint64_t eventnumber);     // Called every event.
 		jerror_t erun();                                 // Called everytime run number changes, provided brun has been called.
 		jerror_t fini();                                 // Called after last event of last event source has been processed.
